Use a h2 section instead of a div
This is very common problem so I think it deserve a full section.
People are new to Tails at this point and they need guidance on how to
check whether the Wi-Fi is working: how to connect to a network.

Also, a div implies HTML and that's more painful for translators.