Add crypttab entry for TailsData_unlocked (refs: #11529)
With this entry we specify the device mapper name of the unlocked
persistent volume. This fixes that the device mapper name is different
when the volume was just created (the name was automatically set to
luks-$UUID by udisks, no way to configure this) and when the volume
was unlocked by the greeter (we specify the name "TailsData_unlocked"
in the cryptsetup command).

The "_unlocked" suffix is required by live-persist, so this is a
requirement for immediately enabling the persistent partition after
creating it.

The "noauto" option ensures that the entry is ignored by
cryptsetup.target, which would otherwise try to unlock the partition
during boot.