Reduce bytes stored in kernel command-line seed (refs: #11897)
By repeatedly dividing the length of the random seed string by two, the
first length that didn't break the syslinux menu was 32, i.e. 16 bytes.

We could check if some length between 32 and 64 would also work. Or we
could use base64 to produce a string that encodes more bytes per
character. But we agreed that we don't want to spend more time on this,
because we also use the random seed stored in LBA 34, and once we
migrate to GRUB, we can probably use a 512 Byte seed by storing it in a
variable.