Add procfs_priv to allow processes to see pids
If hidepid=2 is set when mounting /proc users can only see their own
processes /proc/<pid>. Some processes like systemd-logind and polkit
need to see pids. The users under which this processes are running can
be added to this new procfs_priv group. Using a mount option for /proc this group can later be allowed to access the necessary information.

The GID 151 is choosen > 150 to avoid conflicts with existing GIDs