Greeter: Refactor: Redesign how settings are saved and loaded
Settings are now saved to file immediately when the user changes them,
instead of before login. Beside making the code simpler, this makes it
easier to handle saving the admin password file (which we want to delete
if the user disabled the feature after loading a persistent password
file, but want to overwrite or leave untouched in other cases).