Remove rngd and don't trust the CPU's hwrng (refs: #17154)
Don't trust hardware random number generators, because we don't know if
they are correctly implemented and free of backdoors.

The entropy of (most) hwrngs is still mixed into the entropy pool by
the kernel, but it is not credited as trusted entropy.

Since Linux 5.6, on x86, the kernel doesn't block anymore when reading
from /dev/random, but uses its built-in random number generator, so we
don't have to worry about userspace being blocked when not trusting
hwrngs.