Ensure we synchronize APT data at shutdown before the new persistent partition is unmounted
When testing manually on Bullseye, after enabling Additional Software for
a newly installed package and shutting down, when I inspect the persistent
filesystem from outside of Tails:

* according to e2fsck the filesystem is clean
* live-additional-software.conf LGTM
* persistence.conf LGTM
* the apt directory exists and is empty
* the apt_sync_started flag file exists

⇒ tails-synchronize-data-to-new-persistent-volume-on-shutdown.service started to
do its job but failed to complete it in time.

I suspect that systemd might have unmounted
/media/tails-persistence-setup/TailsData before this service had time to finish
its job, so let's order these 2 units to ensure this one is stopped before
the other.

refs #18839