Revert "Test suite: switch to virtio graphics"
This reverts commit f6f1c927e9219032cb72045217771b7a8b3cb316.

We can do this because the problem that commit fixed does not exist on current
sid anymore. Besides, virtio-gpu/VirGL is not useful in the context of our CI:
its main goal is to "allow the guest operating system to use the capabilities of
the host GPU to accelerate 3D rendering", which is not going to work with nested
virtualization on our CI (even if we enabled the bits that make virtio-gpu
useful in other cases, which we don't).

Let's see if this helps for tails/tails#18953.